Correction message d'erreur quand tentative de suppression d'un compte encore associé à une catégorie
FossilOrigin-Name: a78a32f1b955ed937f8d000c921282dc3fb9e34f

+++ src/include/class.compta_comptes.php
@@ -136,6 +136,16 @@ class Compta_Comptes
             throw new UserException('Ce compte ne peut être supprimé car des opérations comptables y sont liées.');
         }
 
+        if ($db->simpleQuerySingle('SELECT 1 FROM compta_comptes_bancaires WHERE compte = ? LIMIT 1;', false, $id, $id))
+        {
+            throw new UserException('Ce compte ne peut être supprimé car il est lié à un compte bancaire.');
+        }
+
+        if ($db->simpleQuerySingle('SELECT 1 FROM compta_categories WHERE compte = ? LIMIT 1;', false, $id, $id))
+        {
+            throw new UserException('Ce compte ne peut être supprimé car des catégories y sont liées.');
+        }
+
         $db->simpleExec('DELETE FROM compta_comptes WHERE id = ?;', trim($id));
 
         return true;
